move stuff to utils file
