+++ src/screens/Schedule/ScheduleScreen.tsx
@@ -29,7 +29,10 @@ import {
 import { Colors } from 'assets/Colors';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { Platform, View } from 'react-native';
-import { convertCamelToTitleCase } from 'utils/utils';
+import {
+  convertCamelToTitleCase,
+  convertDateObjectToString,
+} from 'utils/utils';
 
 // only display links for the client's approved cases
 // display client's upcoming appointments
@@ -41,29 +44,6 @@ const ScheduleScreen = () => {
   const [calendlyLinks, setCalendlyLinks] = useState<CalendlyLink[]>();
   const [appointments, setAppointments] = useState<Appointment[]>();
   const [switchPage, setSwitchPage] = React.useState(0);
-  const months = [
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'August',
-    'September',
-    'October',
-    'November',
-    'December',
-  ];
-  const daysOfWeek = [
-    'Sunday',
-    'Monday',
-    'Tuesday',
-    'Wednesday',
-    'Thursday',
-    'Friday',
-    'Saturday',
-  ];
 
   useEffect(() => {
     async function loadLinksAndAppointments() {
@@ -135,18 +115,6 @@ const ScheduleScreen = () => {
     return 'Schedule appointments with attorney(s) for newly approved case(s) here. Clicking the button for a specific case will take you to Calendly to choose from your attorney’s availabilities.';
   };
 
-  // convert Date object (appointment.startTime) into a readable string
-  const getDateString = (date: Date) => {
-    const localeTimeString: string = date.toLocaleTimeString();
-    const time: string =
-      localeTimeString.substring(0, localeTimeString.lastIndexOf(':')) +
-      localeTimeString.substring(localeTimeString.lastIndexOf(':') + 3);
-
-    return `${daysOfWeek[date.getDay()]}, ${
-      months[date.getMonth()]
-    } ${date.getDate()} at ${time}`;
-  };
-
   const getUpcomingBody = () => {
     if (appointments === undefined || appointments.length === 0) {
       // if no upcoming appointments
@@ -169,7 +137,7 @@ const ScheduleScreen = () => {
             <AppointmentTextContainer>
               <TextRegular>You have an appointment scheduled for </TextRegular>
               <TextRegularBold>
-                {getDateString(appointment.startTime)}
+                {convertDateObjectToString(appointment.startTime)}
               </TextRegularBold>
               <TextRegular>.</TextRegular>
             </AppointmentTextContainer>
