+++ src/utils/utils.tsx
@@ -1,3 +1,27 @@
+const months = [
+  'January',
+  'February',
+  'March',
+  'April',
+  'May',
+  'June',
+  'July',
+  'August',
+  'September',
+  'October',
+  'November',
+  'December',
+];
+const daysOfWeek = [
+  'Sunday',
+  'Monday',
+  'Tuesday',
+  'Wednesday',
+  'Thursday',
+  'Friday',
+  'Saturday',
+];
+
 // convert camelCase to Title Case
 export const convertCamelToTitleCase = (str: string) => {
   let result = str.replace(/([A-Z])/g, ' $1');
@@ -5,4 +29,14 @@ export const convertCamelToTitleCase = (str: string) => {
   return result;
 };
 
-export default convertCamelToTitleCase;
+// convert a Date object into a readable string
+export const convertDateObjectToString = (date: Date) => {
+  const localeTimeString: string = date.toLocaleTimeString();
+  const time: string =
+    localeTimeString.substring(0, localeTimeString.lastIndexOf(':')) +
+    localeTimeString.substring(localeTimeString.lastIndexOf(':') + 3);
+
+  return `${daysOfWeek[date.getDay()]}, ${
+    months[date.getMonth()]
+  } ${date.getDate()} at ${time}`;
+};
